Added MCMC sampling and tuning codes
Added a code to sample various known density functions using a Metropolis-Hastings MCMC algorithm, and another code to illustrate the process of tuning an MCMC algorithm. In the latter, two flavors of MCMC are tuned, Metropolis-Hastings and stretch-move, which is an ensemble method. Also made a few minor changes in README.md files, comments, and plotting routines.